Profiling Infrastructure (#354)
* Simplify run_single_evaluation

This de-duplicates some duplicated code paths. This makes it easier to
patch profiling calls into the function later on.

* add gpu runtime info to SystemInfo

This way we can tell whether we are using CUDA or ROCm later on.

This also fixes the ROCm fallback path.

* add 'link' report result type

This will be used to communicate external download links
such as profiling results.

* add profiling data infrastructure

A new ProfileResult type is added to run_eval, which is
is returned in the EvalResult type. Among other fields,
this contains the `download_url` field which should be
used by the user to download profiling data. Note that
the actual public download link may not be known in
run_eval.py. In this case, it is the intention that the
launcher fixes up the `download_url` before returning the
results back to libkernelbot.

* github launcher: separate artifact downloading from indexing

The new function `GitHubRun.get_artifact_index` returns a
dict of artifacts available from the run. For each artifact,
the GitHub API URL and public download URL are returned.

The latter is not available directly from the GitHub API,
however, it can be easily constructed from the data that is
available in the worflow result.

`download_artifacts` is replaced by a function which downloads
a specific artifact rather than all of them. Additionally, the
function no longer writes to a temp file when downloading the
artifact; the results of the download request can be piped
directly into zipfile using BytesIO.

* github runner: yield 'profile_data/*' from job as profile data

The idea is that eval_run.py places profiling data in the
profile_data/ directory, which is then automatically exported
to the user. This is done by uploading that directory as the
'profile-data' artifact, then fetching its public download
link and returning that as the ProfileResult.download_url.